trying to reduce "err" noise by passing an err vec
Instead of passing an unbounded additional amount of thickness or width
to the stock creation function, pass a vec of three values. Each value
in the vec is multiplied by `err` and added to its axis ([x,y,z] ==
[length, thickness, width]). If the value in the vec is 2, also shift
the stock -err in that axis, so that the additional measure sticks out
one err on each side.

This seems like it effectively removes a lot of +/-err usage in the
code. Hopefully it's not too opaque in meaning.