begin writing step-by-step build instructions
It seems like this has to start with rethinking file organization.

Step one is to convert all global variables (input and computed) to
functions. This allows their values to be used in files that `use` the
model instead of `include` it.

Being able to use the values from other files is necessary, because for
step-by-step we're gong to want more close-up diagrams, with different
labeling.

Moving labeling to view files means that the labeling isn't created
along with the component, as the README suggested. This is a bit of a
bummer, but I think I may have been wrong that the labeling wouldn't
create too much noise in the model file. As I write the step-by-step,
I'm finding many labels missing, so there would have been even more
content there.

Being able to use the parameter values also means that we can
auto-generate a sed replacement (via a hop through an auto-generated
SCAD file that ECHOs all of their values to the log) that we can apply
to the HTML file(s) to put values in the text.