move shoe rack input params to a separate file, and back to variables
Making input params be variables means they can be overridden from the
command line with -Dparam=value, which opens the possibility for
on-demand re-rendering for different specs.

If the params are going to be variables, they have to be in a file that
can be sourced via `include` instead of `use`, so move them to
a new params.scad.

Computed params are left in the model, as functions, since they should
not be overridden from the commandline.