watermark images
This is an attempt to fight the problem that Pinterest introduces to the
DIY community, where Google turns up exactly the image you're looking
for, and it's a link to Pinterest with absolutely no hope of finding the
original. Maybe this won't be effective, but it's a stab at attaching
some sort of link that may help someone get back here to find the rest
of the plans.

Includes small changes to move some diagrams out from under the watermark.

Also combines all gif creation into one command, which saves some
processing time.