define upper_window_position as a function
Functions and modules are two different namespaces in OpenSCAD, and as
far as I can tell, the syntax makes it unambigous which one is meant.
So, I'm trying out exploiting that here to make it possible to translate
the upper window cutout diagram labels on a mirrored side panel.

I can't use the mirror module here, because I don't want to flip the
labels (which would flip their text). I just want to be able to move
everything the same amount in the opposite direction.

So, I'm adding a file that I've been playing with that is function
definitions of some of the OpenSCAD translation models. What these do is
equivalent of their same-name module, but they operate on vec3 values
instead of scene nodes.

The new toddler tower code uses the `math_funcs` `scale` function to
reverse the direction of translation on the Y axis, to match the left
side panel in the diagram, which was mirrored on Y.