move all animation to anim file
This makes the animation file a bit more complex, but majorly simplifies
the main model file. Since animation is only relevant to that one file,
I think it make sense.

This commit also changes the way the positional shifts are composed
during the animation. Basically, elements are put in their start place,
and then shifts are applied, instead of computing the entire position at
once. Maybe this makes the animation components a little clearer?