make common/squareStock take a vector
I wrote squareStock before I was familiar with how parameters are
commonly passed in OpenSCAD, and before I understood how the scene was
oriented. It's no longer necessary to separate and rename [x,y,z].

Also separating our the "errs" calculation, so it can be used easily
outside of squareStock.

Also changing the default `err` to a special variable `$err` so it can
be overridden in-context when necessary. (I'm prototyping a much larger
project, where rendering conflicts seem to be present at distant zoom
levels.)

Also adds "vabs", which is just "abs", but for each element in a vector.