convert signal-tower to parameterizing component sizes instead of assembly size
While it's neat to say "I want a tower this tall by this wide" and have
a model created to that spec, I found it more useful to say, "I can get
leg material this long, and this angle looks safe, how big would that
tower be?" So, I'm changing to that.

This also allows for easier recording of what I actually built.