fix small-screen hiding menu link clicks
Clicking links in the hiding menu on small screens didn't work - the
menu would just hide, without navigating. This change teaches the
handler that clicks on elements with parents that have "pure-menu..."
classes are not hide clicks.

The check has to be on the element's parent, because not all
sub-elements (for instance, a logo) can have pure-menu classes (it
scretches images).